test(gq): register every .gqt case as its own test in a dedicated corpus crate
