+++ crates/omnigraph-gqt/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "omnigraph-gqt"
+version = "0.10.0"
+edition = "2024"
+description = "GQ logic-test corpus (`cases/*.gqt`) and its runner; one named test per case."
+license = "MIT"
+repository = "https://github.com/ModernRelay/omnigraph"
+homepage = "https://github.com/ModernRelay/omnigraph"
+publish = false
+
+# No doc examples; skip the empty doctest target on every `cargo test -p`.
+[lib]
+doctest = false
+
+[dependencies]
+futures = { workspace = true }
+omnigraph = { package = "omnigraph-engine", path = "../omnigraph", version = "0.10.0" }
+omnigraph-compiler = { path = "../omnigraph-compiler", version = "0.10.0" }
+serde_json = { workspace = true }
+tempfile = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+datatest-stable = "0.3"
+
+# One libtest test per `cases/*.gqt`; mechanism and flags: tests/gq_logic_tests.rs.
+[[test]]
+name = "gq_logic_tests"
+harness = false
+
+[lints]
+workspace = true
