+++ crates/omnigraph-gqt/cases/issue_603_negation_string_predicate_outer_var.gqt
