+++ crates/omnigraph-gqt/cases/mutation_error_typed.gqt
