+++ crates/omnigraph-gqt/cases/order_clause_aggregate_refused.gqt
