+++ crates/omnigraph-gqt/cases/restart_survives_reopen.gqt
